Add Updated Set Logic
