+++ compiler/coenobita-bin/src/callbacks.rs
@@ -1,4 +1,4 @@
-use coenobita_check::{Check, Checker};
+use coenobita_check::{Checker};
 use coenobita_middle::{flow::FlowPair, provenance::ProvenancePair};
 use rustc_driver::{Callbacks, Compilation};
 use rustc_hir::{intravisit::Visitor, Item};
