+++ rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "nightly-2025-02-19"
+profile = "default"
+components = ["rustc-dev", "rustfmt", "clippy", "llvm-tools"]
